plane action based on velocity angle
the sprite action of the planes are now independent of each other, and
determined by the angle of the velocity vector. The hope is that once
the turning mechanism is added, the plane will barrel roll as the
movement angle changes.
Also adding a class for keeping the time elapsed between each frame
outside of the main function.
There's also some few code changes with no visible impact