draw the enemy's plane on the screen
now the player will be able to see the enemy's plane on his side of the
screen, as you'd expect from a multiplayer game. There's a big problem
with this tho: the enemy's plane may be drawn outside the player's
screen, including the enemy's screen.

Also, I added the verification for the planes not to go bellow the floor
in the main function, as it seems it doesn't belong in the game_entity
class.